requirements.txt
Flask==2.2.2
flask-mongoengine==1.0.0
Flask-JWT-Extended==4.4.4
requests==2.28.2
Flask-DotEnv==0.1.2